Create Chewy cookies webpage and provide corresponding link on main page
Added text to main page which reads the name of new cookie recipe no link
as of yet.Added boilerplate html code to cookie webpage along with a
specified title and header to distinguish it from other recipes. Images
repo also was inserted with a new image of chewy cookies that will
have future use in chew cookies webpage.